Update State Methods
Updated state methods for creating and organizing the ingredients list. Method: add_ingts was fixed to allow it to accurately produce the whole title of an ingredient as opposed to just the first word, and allows it to set the ingredient object amount accurately based on the amount of the particular meal it came from.

Method: organize_ingt_list has part of its logic that is not being accessed when run, need to assess and solve. This is the part that senses when there are two ingredient objects in the list that are the same, and turns them into one object after updating the amount within the object.